DBI methods

Implementations of pure virtual functions defined in the DBI package.

# S4 method for bq_dataset
dbConnect(drv, ...)

# S4 method for BigQueryDriver
show(object)

# S4 method for BigQueryDriver
dbGetInfo(dbObj, ...)

# S4 method for BigQueryDriver
dbIsValid(dbObj, ...)

# S4 method for BigQueryDriver
dbDataType(dbObj, obj, ...)

# S4 method for BigQueryConnection
show(object)

# S4 method for BigQueryConnection
dbIsValid(dbObj, ...)

# S4 method for BigQueryConnection
dbDisconnect(conn, ...)

# S4 method for BigQueryConnection,character
dbSendQuery(conn, statement, ...)

# S4 method for BigQueryConnection,character
dbQuoteString(conn, x, ...)

# S4 method for BigQueryConnection,SQL
dbQuoteString(conn, x, ...)

# S4 method for BigQueryConnection,character
dbQuoteIdentifier(conn, x, ...)

# S4 method for BigQueryConnection,SQL
dbQuoteIdentifier(conn, x, ...)

# S4 method for BigQueryConnection
dbDataType(dbObj, obj, ...)

# S4 method for BigQueryConnection,character,data.frame
dbWriteTable(
  conn,
  name,
  value,
  overwrite = FALSE,
  append = FALSE,
  ...,
  field.types = NULL,
  temporary = FALSE,
  row.names = NA
)

# S4 method for BigQueryConnection,character
dbReadTable(conn, name, ...)

# S4 method for BigQueryConnection
dbListTables(conn, ...)

# S4 method for BigQueryConnection,character
dbExistsTable(conn, name, ...)

# S4 method for BigQueryConnection,character
dbListFields(conn, name, ...)

# S4 method for BigQueryConnection,character
dbRemoveTable(conn, name, ...)

# S4 method for BigQueryConnection
dbGetInfo(dbObj, ...)

# S4 method for BigQueryConnection
dbBegin(conn, ...)

# S4 method for BigQueryConnection
dbCommit(conn, ...)

# S4 method for BigQueryConnection
dbRollback(conn, ...)

# S4 method for BigQueryResult
show(object)

# S4 method for BigQueryResult
dbIsValid(dbObj, ...)

# S4 method for BigQueryResult
dbClearResult(res, ...)

# S4 method for BigQueryResult
dbFetch(res, n = -1, ...)

# S4 method for BigQueryResult
dbHasCompleted(res, ...)

# S4 method for BigQueryResult
dbGetStatement(res, ...)

# S4 method for BigQueryResult
dbColumnInfo(res, ...)

# S4 method for BigQueryResult
dbGetRowCount(res, ...)

# S4 method for BigQueryResult
dbGetRowsAffected(res, ...)

# S4 method for BigQueryResult
dbBind(res, params, ...)

Arguments

...

Other arguments to methods.

object

Any R object

dbObj

An object inheriting from DBIObject, i.e. DBIDriver, DBIConnection, or a DBIResult

obj

An R object whose SQL type we want to determine.

conn

A DBIConnection object, as returned by dbConnect().

statement

a character string containing SQL.

x

A character vector to quote as string.

name

The table name, passed on to dbQuoteIdentifier(). Options are:

a character string with the unquoted DBMS table name, e.g. "table_name",

a call to Id() with components to the fully qualified table name, e.g. Id(schema = "my_schema", table = "table_name")

a call to SQL() with the quoted and fully qualified table name given verbatim, e.g. SQL('"my_schema"."table_name"')

value

a data.frame (or coercible to data.frame).

overwrite

a logical specifying whether to overwrite an existing table or not. Its default is FALSE.

append

a logical specifying whether to append to an existing table in the DBMS. Its default is FALSE.

field.types, temporary

Ignored. Included for compatibility with generic.

row.names

A logical specifying whether the row.names should be output to the output DBMS table; if TRUE, an extra field whose name will be whatever the R identifier "row.names" maps to the DBMS (see DBI::make.db.names()). If NA will add rows names if they are characters, otherwise will ignore.

res

An object inheriting from DBIResult.

n

maximum number of records to retrieve per fetch. Use n = -1 or n = Inf to retrieve all pending records. Some implementations may recognize other special values.

params

A list of bindings, named or unnamed.