Deprecated functions for access credentials

bigrquery's main functions for managing credentials are now bq_auth() and bq_auth_configure(). The functions documented here are now deprecated and will eventually be removed.

get_access_cred()

set_access_cred(value)

reset_access_cred()

has_access_cred()

set_oauth2.0_cred(app)

set_service_token(service_token)

Arguments

value
New access credentials, as returned by httr::oauth2.0_token().

app
A Google OAuth application created with httr::oauth_app().

service_token
A JSON string, URL or file, providing a service account token.

Value

get_access_cred() returns the current OAuth2 credentials token.