bigrquery: An Interface to Google's 'BigQuery' 'API'

Easily talk to Google's 'BigQuery' database from R.

Package options

bigrquery.quiet

Verbose output during processing? The default value, NA, turns on verbose output for queries that run longer than two seconds. Use FALSE for immediate verbose output, TRUE for quiet operation.

bigrquery.page.size

Default page size for fetching data, defaults to 1e4.

See also

Author

Maintainer: Hadley Wickham hadley@rstudio.com (ORCID)

Authors:

Jennifer Bryan jenny@rstudio.com (ORCID)

Other contributors:

Kungliga Tekniska Högskolan (strptime implementation) [contributor]

The NetBSD Foundation, Inc. (gmtime implementation) [contributor]

RStudio [copyright holder, funder]