Create a new query job deprecated

Please use api-job and api-perform instead.

insert_query_job(
  query,
  project,
  destination_table = NULL,
  default_dataset = NULL,
  create_disposition = "CREATE_IF_NEEDED",
  write_disposition = "WRITE_EMPTY",
  use_legacy_sql = TRUE,
  ...
)

Arguments

query

SQL query string

destination_table

(optional) destination table for large queries, either as a string in the format used by BigQuery, or as a list with project_id, dataset_id, and table_id entries

default_dataset

(optional) default dataset for any table references in query, either as a string in the format used by BigQuery or as a list with project_id and dataset_id entries

create_disposition

behavior for table creation. defaults to "CREATE_IF_NEEDED", the only other supported value is "CREATE_NEVER"; see the API documentation for more information

write_disposition

behavior for writing data. defaults to "WRITE_EMPTY", other possible values are "WRITE_TRUNCATE" and "WRITE_APPEND"; see the API documentation for more information

use_legacy_sql

(optional) set to FALSE to enable BigQuery's standard SQL.

Value

a job resource list, as documented at https://cloud.google.com/bigquery/docs/reference/rest/v2/jobs

See also