Edit and view auth configuration

These functions give more control over and visibility into the auth configuration than bq_auth() does. bq_auth_configure() lets the user specify their own:

OAuth app, which is used when obtaining a user token. See the vignette How to get your own API credentials for more. If the user does not configure these settings, internal defaults are used. bq_oauth_app() retrieves the currently configured OAuth app.

bq_auth_configure(app, path)

bq_oauth_app()

Arguments

app
OAuth app, in the sense of httr::oauth_app().

path
JSON downloaded from Google Cloud Platform Console, containing a client id (aka key) and secret, in one of the forms supported for the txt argument of jsonlite::fromJSON() (typically, a file path or JSON string).

Value

bq_auth_configure(): An object of R6 class gargle::AuthState, invisibly.

bq_oauth_app(): the current user-configured httr::oauth_app().

See also

Examples

# see the current user-configured OAuth app (probaby `NULL`)
bq_oauth_app()

#> NULL

if (require(httr)) {

  # store current state, so we can restore
  original_app <- bq_oauth_app()

  # bring your own app via client id (aka key) and secret
  google_app <- httr::oauth_app(
    "my-awesome-google-api-wrapping-package",
    key = "123456789.apps.googleusercontent.com",
    secret = "abcdefghijklmnopqrstuvwxyz"
  )
  bq_auth_configure(app = google_app)

  # confirm current app
  bq_oauth_app()

  # restore original state
  bq_auth_configure(app = original_app)
  bq_oauth_app()
}

#> Loading required package: httr

#> NULL

if (FALSE) {
# bring your own app via JSON downloaded from GCP Console
bq_auth_configure(
  path = "/path/to/the/JSON/you/downloaded/from/gcp/console.json"
)
}